implement a consistent api (resolves #74) and a consistent folder structure (resolves #75)
