+++ docs/package.json
@@ -2,7 +2,7 @@
   "name": "quantum-site",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
     "start": "quantum watch"
   },
