+++ quantum-changelog/lib/config.js
@@ -1,3 +1,4 @@
+'use-strict'
 const path = require('path')
 
 function defaultDocsUrlLookup (version, api) {
