+++ quantum-code-highlight/lib/quantum-syntax.js
@@ -1,3 +1,4 @@
+'use-strict'
 module.exports = function (hljs) {
   const INLINE_CONTENT = {
     className: 'subst',
