+++ quantum-diagram/example/package.json
@@ -2,7 +2,7 @@
   "name": "quantum-diagram-example",
   "version": "0.1.0",
   "description": "Example for quantum-diagram",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
     "start": "node build.js"
   },
