+++ quantum-dom/package.json
@@ -15,14 +15,14 @@
     "url": "https://github.com/ocadotechnology/quantumjs/issues"
   },
   "homepage": "https://quantum.hexagonjs.io",
-  "main": "index.js",
+  "main": "lib/index.js",
   "directories": {
     "test": "test"
   },
   "scripts": {
     "test": "./node_modules/istanbul/lib/cli.js cover --dir coverage ./node_modules/mocha/bin/_mocha -- --ui bdd -R spec test && npm run lint",
-    "lint": "eslint index.js test/*.js",
-    "fix-lint-issues": "eslint --fix index.js test/*.js"
+    "lint": "eslint lib/*.js test/*.js",
+    "fix-lint-issues": "eslint --fix lib/*.js test/*.js"
   },
   "config": {
     "blanket": {
