+++ quantum-dom/test/spec.js
@@ -1,3 +1,4 @@
+'use-strict'
 const chai = require('chai')
 const chaiAsPromised = require('chai-as-promised')
 const dom = require('..')
