+++ quantum-js/lib/json.js
@@ -1,3 +1,4 @@
+'use-strict'
 /*
 
      ____                    __                      _
