+++ quantum-js/lib/read.js
@@ -1,3 +1,4 @@
+'use-strict'
 /*
 
      ____                    __                      _
