+++ quantum-js/lib/watch.js
@@ -1,3 +1,4 @@
+'use-strict'
 /*
      ____                    __                      _
     / __ \__  ______ _____  / /___  ______ ___      (_)____
