+++ quantum-js/test/page.spec.js
@@ -1,3 +1,4 @@
+'use-strict'
 require('chai').should()
 
 const quantum = require('..')
