+++ quantum-js/test/parse.spec.js
@@ -1,3 +1,4 @@
+'use-strict'
 const chai = require('chai')
 chai.should()
