+++ quantum-js/test/select.spec.js
@@ -1,3 +1,4 @@
+'use-strict'
 const chai = require('chai')
 const should = chai.should()
 const expect = chai.expect
