+++ quantum-js/test/stringify.spec.js
@@ -1,3 +1,4 @@
+'use-strict'
 const Promise = require('bluebird')
 const fs = Promise.promisifyAll(require('fs-extra'))
 const quantum = require('..')
