+++ quantum-js/test/watch.spec.js
@@ -1,3 +1,4 @@
+'use-strict'
 const should = require('chai').should()
 const Promise = require('bluebird')
 const fs = Promise.promisifyAll(require('fs-extra'))
